Add function to hide empty items grid on removal
Introduce `checkAndHideEmptyItemsGrid` function to check
if the items grid is empty and hide it accordingly. This
function is added to `PageEdit.cshtml` and the public API
of `SectionModal` in `Content.Page.Section.Modal.js`.

Update `removeItem` to call this function after item
removal, ensuring the UI accurately reflects the current
state of the items grid.